Add native es-VE
hii! The reason why i'm adding this is because it's essential to make the content feel more natural and relatable for users from another countries. As you may know the Spanish language varies significantly across different regions, and oobviously adapting it to the local culture enhances a lot the user experience, this includes specific vocabulary and expressions that come from Venezuela itself, distinguishing it from other regions that also speak Spanish in the world. 🤗